Bonus: break the circular dependency
ItemList.svelte -> iOSThirdPartyData.svelte -> ItemList.svelte
